% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{areFilesUTF8}
\alias{areFilesUTF8}
\title{areFilesUTF8}
\usage{
areFilesUTF8(filePaths)
}
\arguments{
\item{filePaths}{A character vector of file paths.}
}
\value{
Logical values
}
\description{
Generalization of \code{ospsuite.utils::isFileUTF8()} to check multiple files.
}
\keyword{internal}
