% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-engine-ui.R
\name{copyQualificationTemplate}
\alias{copyQualificationTemplate}
\title{copyQualificationTemplate}
\usage{
copyQualificationTemplate()
}
\description{
Copy and open the \code{{ospsuite.reportingengine}} template
for qualification workflows in current project directory.
}
