% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-ui.R
\name{displayBranchNames}
\alias{displayBranchNames}
\title{displayBranchNames}
\usage{
displayBranchNames(tree)
}
\arguments{
\item{tree}{A list of simulation paths}
}
\value{
A list with branch names to display
}
\description{
Display branch names of a simulation tree
}
\keyword{internal}
