% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-ui.R
\name{formatSimulationTree}
\alias{formatSimulationTree}
\title{formatSimulationTree}
\usage{
formatSimulationTree(simulation)
}
\arguments{
\item{simulation}{A \code{Simulation} object}
}
\value{
A list for the simulation tree.
}
\description{
Format the simulation tree to be displayed by \code{{shinyTree}} package
}
\keyword{internal}
