% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-ui.R
\name{insertOSPSuiteTemplate}
\alias{insertOSPSuiteTemplate}
\title{insertOSPSuiteTemplate}
\usage{
insertOSPSuiteTemplate()
}
\description{
Insert template code in current document aiming at
providing a starting point for users to interact with the \code{{ospsuite}} package.
The code will include the following steps:
\itemize{
\item loads a simulation
\item update a simulation parameter
\item add output paths to a simulation
\item run a simulation
\item format simulation results
\item save simulation results
\item run a PK analysis
\item format PK analysis results
\item save PK analysis results
}
}
