% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-ui.R
\name{runSimulationTree}
\alias{runSimulationTree}
\title{runSimulationTree}
\usage{
runSimulationTree()
}
\value{
A shiny app
}
\description{
Popup shiny app that displays a simulation tree.
Selected branches are copied to clipboard when closing app.
}
