% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-tables.R
\name{tableOfPKParameters}
\alias{tableOfPKParameters}
\title{tableOfPKParameters}
\usage{
tableOfPKParameters()
}
\value{
A datatable of available PK Parameters
}
\description{
This function will display a datatable of available PK Parameters
}
