% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-tables.R
\name{tableOfUnits}
\alias{tableOfUnits}
\title{tableOfUnits}
\usage{
tableOfUnits()
}
\value{
A datatable of available OSPSuite units and their corresponding dimensions.
}
\description{
This function will display a datatable of available OSPSuite units and their corresponding dimensions.
}
