% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{utfCheck}
\alias{utfCheck}
\title{utfCheck}
\usage{
utfCheck()
}
\value{
A shiny app
}
\description{
Popup shiny app that displays the list of files in a selected folder
along with a check if the file is UTF-8 encoded.
Selected branches are copied to clipboard when closing app.
}
