% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-engine-ui.R
\name{viewReport}
\alias{viewReport}
\title{viewReport}
\usage{
viewReport(filePath = NULL)
}
\arguments{
\item{filePath}{The path to the report file. If \code{NULL}, a file dialog will be opened.}
}
\value{
filePath The path to the report file.
}
\description{
Open a markdown or docx report from \code{{ospsuite.reportingengine}} from user default software.
}
